+++ packages/directives/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @viaz/directives
 
+## 0.0.2-beta.11
+
+### Patch Changes
+
+- add from-table event
+- Updated dependencies
+  - @viaz/types@0.0.2-beta.11
+
 ## 0.0.2-beta.10
 
 ### Patch Changes