+++ packages/enums/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @imsjs/ims-ui-enums
 
+## 0.0.2-beta.11
+
+### Patch Changes
+
+- add from-table event
+- Updated dependencies
+  - @viaz/hooks@0.0.2-beta.11
+
 ## 0.0.2-beta.10
 
 ### Patch Changes